Add Rule.Parse and Rule.Syntax modules.
