+++ src/Chromar/Rule/TH.hs
@@ -16,7 +16,9 @@ import "template-haskell" Language.Haskell.TH
 import Language.Haskell.TH.Quote (QuasiQuoter(..))
 import "template-haskell" Language.Haskell.TH.Syntax (showName)
 import Text.ParserCombinators.Parsec (parse)
-import Chromar.MRuleParser (LAgent(..), RAgent(..), ARule(..), SRule(..), parseRule)
+
+import Chromar.Rule.Syntax (LAgent(..), RAgent(..), ARule(..), SRule(..))
+import Chromar.Rule.Parse (parseRule)
 import Chromar.MAttrs (fillAttrs)
 import Chromar.Enriched.Syntax (SEr)
 import qualified Chromar.Enriched.TH as RE (quoteEr)
